Get ready for Kubernetes
Remove the cron daemon and Postfix MTA from this image. Replace the
Postfix MTA with ssmtp to make it much easier to use this image, and to
make it work better with Kubernetes.

The Docker image has been converted to use a specific Github repository
layout to control the project. The README.md file contains full details.